+++ src/modules/capital/capital.service.ts
@@ -107,7 +107,7 @@ export class CapitalService {
             }
           }
           // 不是管理员账号，执行下面
-          const routePermission = routeFind.permission;
+          const routePermission = routeFind.menuPermission;
           const result = await this.menuService.findByPermission(routePermission);
           if (result) {
             return (this.response = {