+++ src/schemas/capital/role.schema.ts
@@ -21,7 +21,10 @@ export class Role extends Document {
   roleType: number; // 角色类型
 
   @Prop()
-  permission: Array<string>; // 角色权限
+  menuPermission: Array<string>; // 角色菜单权限
+
+  @Prop()
+  apiPermission: Array<string>; // 角色api权限
 }
 
 export const RoleSchema = SchemaFactory.createForClass(Role);